+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # dev
 - Prevent too large IAM policy for running tasks
+- Prevent creating unused volume to mount container files
 
 # 3.6.2
 - Fix bad `is_spot` parameter use